+++ resources/release-notes.md
@@ -1,3 +1,7 @@
+## App
+
+- (IMAP) Add option to override figure name for field views
+
 ## Software
 
 - (HelioSwarm) Add calibration for SEM sensors